refactor: rvr 64bit address ptr (#2928)
refactor address pointer related variables to 64bit instead of 32bit.
this PR modifies multiple files:
1) rvr core c and trace headers
2) ffi files
3) rvr code for extensions


reth benchmark result (original 32bit address space variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28121173062
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122726878

reth benchmark result (64bit address pointer variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28119229848
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122157089

By comparing "Exec Median (ms)" columns, it seems that 64bit version is
`10ms` faster or slower compared with the original 32bit version (in
both pure and metered modes). I am not sure why, it could be system
noises affecting it. Also, the comparison only compares execution speed,
excluding other possible important metrics when evaluating this pr.

resolves INT-8321

---------

+++ crates/rvr/rvr-openvm/c/rvr_ext_wrappers.c
@@ -15,34 +15,34 @@
 
 /* ── Memory access (single word) ───────────────────────────────────── */
 
-uint64_t rd_mem_u64_wrapper(RvState* s, uint32_t addr) {
+uint64_t rd_mem_u64_wrapper(RvState* s, uint64_t addr) {
   return rd_mem_u64(s->memory, addr);
 }
-void trace_rd_mem_u64_wrapper(RvState* s, uint32_t addr, uint64_t val) {
+void trace_rd_mem_u64_wrapper(RvState* s, uint64_t addr, uint64_t val) {
   trace_rd_mem_u64(s, addr, val);
 }
 
-void rd_mem_u64_range_wrapper(RvState* s, uint32_t base_addr, uint64_t* out,
+void rd_mem_u64_range_wrapper(RvState* s, uint64_t base_addr, uint64_t* out,
                               uint32_t num_words) {
   rd_mem_u64_range(s, base_addr, out, num_words);
 }
 
-void wr_mem_u64_range_wrapper(RvState* s, uint32_t base_addr,
+void wr_mem_u64_range_wrapper(RvState* s, uint64_t base_addr,
                               const uint64_t* vals, uint32_t num_words) {
   wr_mem_u64_range(s, base_addr, vals, num_words);
 }
 
-void trace_rd_mem_u64_range_wrapper(RvState* s, uint32_t base_addr,
+void trace_rd_mem_u64_range_wrapper(RvState* s, uint64_t base_addr,
                                     const uint64_t* vals, uint32_t num_words) {
   trace_rd_mem_u64_range(s, base_addr, vals, num_words);
 }
 
-void trace_wr_mem_u64_range_wrapper(RvState* s, uint32_t base_addr,
+void trace_wr_mem_u64_range_wrapper(RvState* s, uint64_t base_addr,
                                     const uint64_t* vals, uint32_t num_words) {
   trace_wr_mem_u64_range(s, base_addr, vals, num_words);
 }
 
-void trace_mem_access_u64_range_wrapper(RvState* s, uint32_t base_addr,
+void trace_mem_access_u64_range_wrapper(RvState* s, uint64_t base_addr,
                                         uint32_t num_words,
                                         uint32_t addr_space) {
   trace_mem_access_u64_range(s, base_addr, num_words, addr_space);
